[ndk][conf] Make an llvm-config for cross-compilation
The llvm-config-host program that was installed wasn't
host-related at all; it runs on build and also reports
about build. This program is still needed for the build
to work, so a new BuildToolsCross folder is created and
in there, an llvm-config is built that runs on build
and reports about host.

For this to work, the Makefile that generates llvm-config
and the one that generates BuildVariables.inc must not be
the same, so Makefile.buildvars is used, and an extra var
CROSS_DEPTH is used to enable BuildToolsCross to step one
folder higher up into the top-level build folder to get
the correct values to be emitted into BuildVariables.inc.

The old BuildTools/*/bin/llvm-config is now installed
as llvm-config-build, and the new one is installed as
llvm-config-host

This change has also been submitted to llvm-commits ML:

http://lists.cs.uiuc.edu/pipermail/llvm-commits/Week-of-Mon-20130826/185708.html

Cherry-picked from release_33 branch.

Change-Id: Icc630908044c8bd54166bfbe7295414d7f22793b