[ndk] Add -disable-global-ctor-const-promotion option.
The global variable optimization will try to evaluate the
constant global variable and remove the constructors.

However, some legacy code will try to perform const_cast<>()
and assign to the "constant" to avoid static initialization
order fiasco.  To workaround those old code, this commit
adds a new option to disable the promotion of the global
constant with constructors.

Change-Id: If5146f1ebc9139008b5899cd474c5f57aef0697a